Fix. Change from Integer to Long at celular

+++ src/main/java/dev/luisghtz/platzi_market/percistence/entities/Cliente.java
@@ -23,7 +23,7 @@ public class Cliente {
 
   private String apellidos;
 
-  private Integer celular;
+  private Long celular;
 
   private String direccion;
